Split event catering page into components
